Modèle de base des fiches (DB + Backend) Service de gestion des fiches (Backend) WS de gestion loggué des fiches (Backend) WS de visionnage publique des fiches (Backend)

+++ src/main/java/be/arbiter/clouddragonsheet/data/entities/BaseEntity.java
@@ -1,15 +1,12 @@
 package be.arbiter.clouddragonsheet.data.entities;
 
-import jakarta.persistence.*;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.util.StringUtils;
-
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.stream.Collectors;
 @MappedSuperclass
 @Data
 @AllArgsConstructor
@@ -17,5 +14,5 @@ import java.util.stream.Collectors;
 public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 }
